Minute csv data strategy working.
Currently there is the ability to upload csv data. Look to the dev.py for a strategy to copy and work with. Currently it is set up for a one minute backtesting strategy. The benchmarks in main were turned off for the time being as they conflict with the one minute time frame.

To run a daily test, look in the data directory. There are two daily dataframes to use, DJP and SPY.

Also note how easy it is to add in indicator lines in pandas. I added in SMA15 and SMA 100. This is so easy and we could leverage libraries like bta-lib and finta and ta-lib making technical indicators available immediatly.

The current method of loading works well. I had visions of multi indexing but I will have to see if we need to do that.

So far I've only added one dataframe. Multiple ones will be easy but we need to scrub the index, which I will look at today.