Corrected OCO bug.
I was unable to replicate both oco orders filling in the same bar, however, I'm sure it has to do with threading. Added in an attribute to check off if other order has been filled for any bracket, oto, oco. This will no catch the order at the top of `process_pending_orders` and stop it from being processed, not dependent on the threads.