Updated alpaca_data.py To Work The Same As Other Brokers
Alpaca would previously return fewer number of rows than requested even when it was required. This was a bug recently introduced, which could cause the backtesting to work perfectly but not live as alpaca would not return an appropriate amount of rows in the dataframe. Thus any calculations being done with Alpaca's historical data may not work